Display and edit project long texts in HTML
Previously we have converted all parsed content to markdown and
converted it to HTML when displaying it. This enabled editing the
project description and license text as markdown. However, this approach
doesn't give us the same rendering as the projects have on GitHub, and
makes supporting other input languages supported on GitHub harder (such
as reStructuredText). Converting everything to HTML and then editing the
HTML directly is much easier.