Restart update project info consumer if necessary
The consumer is a long-running PHP process, and due to our use of lazy
services (see config.yml), the DB connection is initiated only once when
the service starts. After the wait_timeout of MySQL the database
connection is terminated and no futher queries are possible. If we only
capture the resulting exception and continue, we'll never get back into
working state with a fresh DB connection.

Reconnecting inside the consumer is rather tricky, so the easiest and
cleanest way is to spawn a new instance of the consumer (systemd takes
care of that).