Allow connecting users to multiple OAuth providers
Currently, we only allowed connecting a LibreCores account to exactly
one GitHub or Google account via OAuth. In addition, the functionality
to connect to such an account after registering an account on LibreCores
was broken.

With this commit, things change in the following way:
- A LibreCores account can be connected to a Google and a GitHub account
  at the same time. (Or to none at all.)
- The groundwork to disconnect from an OAuth account is laid.
- The existing database entries are migrated to the new scheme.
- In contrast to the standard way the HWIOAuthBundle handles OAuth
  providers, we also store the access token. We can use this token to
  register commit hooks and query other non-public information on behalf
  of the user.

Design note: I chose the rather "un-nice" way of adding two columns for
each provider, as it (a) is in line with what standard HWIOAuthBundle
does, thus making understanding for newcomers easier, and (b) we don't
expect many OAuth providers to be added in the future. In this case this
approach has performance benefits.