Vagrant: Restart librecores-rabbitmq with nfs
When loosing the NFS connection, we also loose the configuration file
used by our RabbitMQ handler, causing the startup to fail. If it fails
too often, the service will be disabled and must be restarted manually.

This fix enhances the systemd service file as to always restart our
librecores-rabbitmq service if the NFS client target is restarted.