OAuth: Add failure path to auto-registration
If a new users logs in through OAuth (github/google), a new account is
created with the username given by the OAuth provider and the user can
start using LibreCores instantly. If for some reason this auto-creation
of the user account fails (e.g. if the username is already used, or
incompatible with our rules), the failure path kicks in.

In this path, the user is redirected to the registration page. There
he/she can adjust the username and email address and register a new
account. The remaining data from the OAuth provider is then copied to
the newly created user and it is automatically connected to the OAuth
account.

In consequence, the user was able to choose his/her preferred username
and still log in through OAuth.

Solving this in the FOSUserBundle/HWIOAuth framework is extremely
annoying and took many tries to get right. The final solution is not the
nicest one, but works at least.