Planet: Correctly display text with entities
Our RSS parser (simplepie) returns content with XML entities as they are
present in the feed. Then twig escapes the content again, resulting in
double-escaped content.

For example, we have 'FOSSi Foundation - News &amp; Posts' in our
visible output, or 'FOSSi Foundation - News &amp;amp; Posts' in the
HTML.

This patch removes the entities from all content we get from simplepie.
Proper escaping will be applied safely by twig. This way we can rely on
the rather well-documented behavior of twig, and don't need to trust the
entities to be present in the simplepie content.