Disable Sentry in dev/test environments
As recommended by their documentation
(https://github.com/getsentry/sentry-symfony):

Note that, with this snippet of code, the bundle will be enabled only for
the staging and prod environment; adjust it to your needs. It's discouraged
to enable this bundle in the test environment, because the Sentry client will
change the error handler, which is already used by other packages like
Symfony's deprecation handler (see #46 and #95).