Deployment: Fix RabbitMQ service management
We use one service for each RabbitMQ consumer handling various web
background jobs. For the new notification service to work, it actually
needs to be enabled. And it seems that the GitHub metadata crawler was
never enabled -- and it fact it breaks when enabled. Adding code to
enable it once the underlying bug has been fixed.