Fail gracefully if no user auth is available for GH
The way we get project metadata from GitHub requires an authenticated
user. We get the required token if a user adds a project to LibreCores
through some, but not all mechanisms. Fail more gracefully if we don't
have the required information available.

This, however, doesn't yet give us the ability to get project metadata
without a user access token. This work is tracked in #446.