Wrap all project updates in a DB transaction
The repository crawler and associated mechanics to update a project did
flush the database at various times, but didn't use a single
transaction. This caused problems if a sub-step of the processing
failed. For example, cloc could fail, but the commits were already
extracted from the repository. Since subsequent runs of the crawler only
take changed commits into account, cloc would not run again in the next
crawling run and project metadata would be missing.

Avoid all of this by using a single database transaction, and rolling it
back if anything fails in the processing.

This makes the "inProcessing" field useless (it wouldn't be seen outside
of the transaction). Replace it with a more meaningful updated date for
the license and README fields. For backwards-compat, continue to support
the isInProcessing() getter.