Explicitly set cache directory for HTML Purifier
If this directory is not set explicitly, it ends up being in the vendor
directory, which isn't writable in non-development setups.

The solution chosen here of injecting the Container into the
GitRepoCrawler is probably not the most beautiful solution. A better
solution would be to make the MarkdownToHtml converters services again,
but then the challenge becomes how to get the GitHub owner/repo
information into it in a sane way.