Re-enable custom user form validation
The validation.xml file, which adds our custom user/org name validator,
was not read, since it was in the wrong directory. The Symfony
documentation wasn't overly helpful, but the source code in
FrameworkExtension::registerValidatorMapping() indicates that the
correct directory for Symfony 3.4 is `%kernel.project_dir%/config/validator`.

This commit also removes all duplicated configuration from the
FOSUserBundle validation configuration; it is not possible to override
validator configuration, only to extend it. Having the file contents
duplicated indicates that overriding would be possible.

(Overriding is only possible if a different validation group is chosen.)